simple-meds-backend/requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
google-genai==1.64.0
google-auth-oauthlib==1.0.0
google-api-python-client==2.90.0
pillow
pytesseract==0.3.13
requests==2.31.0
pytz